Remove local API implementation and swap for ApiClient
