--- src/infrastructure/applications/static.js
@@ -1,7 +1,0 @@
-const listAllServices = async () => {
-  return Promise.resolve(null);
-};
-
-module.exports = {
-  listAllServices,
-};
