--- src/infrastructure/organisations/static.js
@@ -1,12 +1,0 @@
-const getOrganisationAndServiceForUserV2 = async () => {
-  return Promise.resolve([]);
-};
-
-const getApproversForOrganisation = async () => {
-  return Promise.resolve();
-};
-
-module.exports = {
-  getOrganisationAndServiceForUserV2,
-  getApproversForOrganisation,
-};
